pyproject.toml
[tool.poetry]
name = "tests-plotter"
version = "0.1.0"
description = ""
authors = ["Nick <tnick1502@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
bcrypt = "3.2.0"
aiobotocore = "^2.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
